CodeChallenge/Repositories/CompensationRepository.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;
using CodeChallenge.Models;
using Microsoft.Extensions.Logging;
using Microsoft.EntityFrameworkCore;
using CodeChallenge.Data;

namespace CodeChallenge.Repositories
{
	public class CompensationRepository : ICompensationRepository
    {
        private readonly CompensationContext _compensationContext;
        private readonly ILogger<ICompensationRepository> _logger;

        public CompensationRepository(ILogger<ICompensationRepository> logger, CompensationContext compensationContext)
		{
            _logger = logger;
            _compensationContext = compensationContext;
		}

        // Add the given Compensation to the Compensation DB
        public Compensation Add(Compensation compensation)
        {
            _compensationContext.Compensations.Add(compensation);
            return compensation;
        }

        // Find the Compensation for the EmployeeId from the DB
        public async Task<Compensation> GetCompensationByEmployeeId(string id)
        {
            return await _compensationContext.Compensations
                .SingleOrDefaultAsync(e => e.EmployeeId == id);
        }

        public Task SaveAsync()
        {
            return _compensationContext.SaveChangesAsync();
        }
    }
}

